Update backport.rb
The require_relative method is not included since it was introduced in Ruby 1.9.2, later than the minimum version (1.8) for which the original code is checked. Also, CGI.escape is used instead of CGI::escape to reflect the changes in the CGI library. I almost needed this for a project on some very old infra. But ended up not needing it and didn't want it to go to waste.