Update fileutils.rb
Here we have a Ruby script that checks if kerberos authentication is enabled or disabled. If it's set to true, the script outputs "Kerberos authentication is enabled". And if it's set to false, the script outputs "Kerberos authentication is disabled".

Along with that, I added some extra methods in the FileUtils module that allow you to set up various configuration options for a WebHDFS client. For example, you can set the host and port of the NameNode, enable or disable SSL and HTTPFS mode, and so on.